Code Improvements
Changed a few things to make it so much better, Slider mapping is now V0:0|V1:10|V2:20|V3:30|V4:40| and that seems to make reading it so much better due to the uniform format preventing misreading of malformed items. Cleaned up sending of items as well to introduce more averaging so it doesn't spam +- 1 of the value over and over
EXE updated as well. Still haven't swapped to miso yet tho but works so much better now so I'm not sold on it yet